requirements.txt
langchain==0.3.9
langchain-core==0.3.19
langgraph==0.2.46
langchain-community==0.3.9
langchain-openai==0.2.5
python-dotenv==1.0.1
pydantic==2.7.1
sqlalchemy==2.0.28
snowflake-connector-python==3.10.1
snowflake-sqlalchemy==1.5.5
pinecone-client==4.1.1
openai==1.51.0
pyyaml==6.0.1
pytest==8.1.1
jupyter==1.0.0
pandas==2.2.1
numpy==1.26.4
